+++ executors/rust/1.3/src/relativedatetime_fmt.rs
@@ -1,195 +1,104 @@
-// https://docs.rs/icu/1.3.2/icu/relativetime/struct.RelativeTimeFormatter.html
+ // https://docs.rs/icu/1.3.2/icu/datetime/struct.DateTimeFormatter.html
 
-use fixed_decimal::FixedDecimal;
-use icu::locid::Locale;
+use icu::calendar::DateTime;
+use icu::calendar::{buddhist::Buddhist,
+                    chinese::Chinese,
+                    coptic::Coptic,
+                    dangi::Dangi,
+                    ethiopian::Ethiopian,
+                    gregory::Gregorian,
+                    hebrew::Hebrew,
+                    indian::Indian,
+                    iso::Iso,
+                    japanese::Japanese,
+                    julian::Julian,
+                    persian::Persian,
+                    roc::Roc,
+                    Date};
+
+use icu::datetime::{options::length, DateTimeFormatter};
+// use icu::datetime::input::DateTimeInput;
 use icu_provider::DataLocale;
 
-use icu::relativetime::{RelativeTimeError, RelativeTimeFormatter, RelativeTimeFormatterOptions};
+use icu::locid::Locale;
 
-use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
+use serde::{Deserialize,Serialize};
 
 #[derive(Deserialize, Serialize, Debug)]
 #[serde(rename_all = "camelCase")]
-struct RelativeDateTimeFormatterOptions {
-    style: Option<String>,
-    numbering_system: Option<String>,
-}
-
-fn get_formatter_from_unit_style(
-    locale: DataLocale,
-    unit: String,
-    style: String,
-    options: RelativeTimeFormatterOptions,
-) -> Result<RelativeTimeFormatter, RelativeTimeError> {
-    if unit == "year" {
-        if style == "long" {
-            RelativeTimeFormatter::try_new_long_year(&locale, options)
-        } else if style == "short" {
-            RelativeTimeFormatter::try_new_short_year(&locale, options)
-        } else if style == "narrow" {
-            RelativeTimeFormatter::try_new_narrow_year(&locale, options)
-        } else {
-            // Assume long
-            RelativeTimeFormatter::try_new_long_year(&locale, options)
-        }
-    } else if unit == "quarter" {
-        if style == "long" {
-            RelativeTimeFormatter::try_new_long_quarter(&locale, options)
-        } else if style == "short" {
-            RelativeTimeFormatter::try_new_short_quarter(&locale, options)
-        } else if style == "narrow" {
-            RelativeTimeFormatter::try_new_narrow_quarter(&locale, options)
-        } else {
-            // Assume long
-            RelativeTimeFormatter::try_new_long_quarter(&locale, options)
-        }
-    } else if unit == "month" {
-        if style == "long" {
-            RelativeTimeFormatter::try_new_long_month(&locale, options)
-        } else if style == "short" {
-            RelativeTimeFormatter::try_new_short_month(&locale, options)
-        } else if style == "narrow" {
-            RelativeTimeFormatter::try_new_narrow_month(&locale, options)
-        } else {
-            // Assume long
-            RelativeTimeFormatter::try_new_long_month(&locale, options)
-        }
-    } else if unit == "week" {
-        if style == "long" {
-            RelativeTimeFormatter::try_new_long_week(&locale, options)
-        } else if style == "short" {
-            RelativeTimeFormatter::try_new_short_week(&locale, options)
-        } else if style == "narrow" {
-            RelativeTimeFormatter::try_new_narrow_week(&locale, options)
-        } else {
-            // Assume long
-            RelativeTimeFormatter::try_new_long_week(&locale, options)
-        }
-    } else if unit == "day" {
-        if style == "long" {
-            RelativeTimeFormatter::try_new_long_day(&locale, options)
-        } else if style == "short" {
-            RelativeTimeFormatter::try_new_short_day(&locale, options)
-        } else if style == "narrow" {
-            RelativeTimeFormatter::try_new_narrow_day(&locale, options)
-        } else {
-            // Assume long
-            RelativeTimeFormatter::try_new_long_day(&locale, options)
-        }
-    } else if unit == "hour" {
-        if style == "long" {
-            RelativeTimeFormatter::try_new_long_hour(&locale, options)
-        } else if style == "short" {
-            RelativeTimeFormatter::try_new_short_hour(&locale, options)
-        } else if style == "narrow" {
-            RelativeTimeFormatter::try_new_narrow_hour(&locale, options)
-        } else {
-            // Assume long
-            RelativeTimeFormatter::try_new_long_hour(&locale, options)
-        }
-    } else if unit == "minute" {
-        if style == "long" {
-            RelativeTimeFormatter::try_new_long_minute(&locale, options)
-        } else if style == "short" {
-            RelativeTimeFormatter::try_new_short_minute(&locale, options)
-        } else if style == "narrow" {
-            RelativeTimeFormatter::try_new_narrow_minute(&locale, options)
-        } else {
-            // Assume long
-            RelativeTimeFormatter::try_new_long_minute(&locale, options)
-        }
-    } else if unit == "second" {
-        if style == "long" {
-            RelativeTimeFormatter::try_new_long_second(&locale, options)
-        } else if style == "short" {
-            RelativeTimeFormatter::try_new_short_second(&locale, options)
-        } else if style == "narrow" {
-            RelativeTimeFormatter::try_new_narrow_second(&locale, options)
-        } else {
-            // Assume long
-            RelativeTimeFormatter::try_new_long_second(&locale, options)
-        }
-    } else {
-        // An unknown unit!
-        RelativeTimeFormatter::try_new_narrow_second(&locale, options)
-    }
+struct RDTFormatOptions {
+    calendar: Option<String>,
+    numbering_system: Option<String>
 }
 
 pub fn run_relativedatetimeformat_test(json_obj: &Value) -> Result<Value, String> {
     let label = &json_obj["label"].as_str().unwrap();
-
+    
     // If there are unsupported values, return
     // "unsupported" rather than an error.
     let options = &json_obj["options"]; // This will be an array.
 
-    let option_struct: RelativeDateTimeFormatterOptions =
-        serde_json::from_str(&options.to_string()).unwrap();
+    let option_struct: RDTFormatOptions = serde_json::from_str(&options.to_string()).unwrap();
+    let numbering_system_str = &option_struct.numbering_system;
 
     let locale_json_str: &str = json_obj["locale"].as_str().unwrap();
-
-    let locale_str: String = locale_json_str.to_string();
-    let lang_id = if let Ok(lc) = locale_str.parse::<Locale>() {
-        lc
-    } else {
-        return Ok(json!({
-            "label": label,
-            "error_detail": {"locale": locale_str},
-            "error_type": "locale problem",
-        }));
-    };
-    let data_locale = DataLocale::from(lang_id);
-
-    let count_str: &str = json_obj["count"].as_str().unwrap();
-    let count = count_str
-        .parse::<FixedDecimal>()
-        .map_err(|e| e.to_string())?;
-
-    let unit: &str = json_obj["unit"].as_str().unwrap();
-
-    let mut style = "";
-    if option_struct.style.is_some() {
-        style = option_struct.style.as_ref().unwrap();
+    let mut locale_str: String = locale_json_str.to_string();
+    if numbering_system_str.is_some() {
+        locale_str = locale_json_str.to_string() + "-u-nu-" +
+            &numbering_system_str.as_ref().unwrap();
     }
 
-    // Update when ICU4X supports non-Latn numbering systems
-    if option_struct.numbering_system.is_some() {
-        let numbering_system = option_struct.numbering_system.as_ref().unwrap();
+    let langid: Locale = json_obj
+        .get("locale")
+        .map(|locale_name| local_str)
+        .unwrap_or_default();
 
-        if numbering_system != "latn" {
-            return Ok(json!({
-                "error": "Number system not supported",
-                "error_msg": numbering_system,
-                "label": label,
-                "unsupported": "non-Latn numbering system not implemented",
-            }));
-        }
-    }
+    let data_locale = DataLocale::from(langid);
 
-    // Use unit & style to select the correct constructor.
-    let relative_time_formatter = get_formatter_from_unit_style(
-        data_locale,
-        unit.to_string(),
-        style.to_string(),
-        RelativeTimeFormatterOptions::default(),
+    // 
+    let mut _unsupported_options: Vec<&str> = Vec::new();
+
+    // handle options - maybe done?
+    
+    // Set up DT options
+    let dt_options = length::Bag::from_date_time_style(
+        length::Date::Medium,
+        length::Time::Short
     );
+    
+    let option_struct: DateTimeFormatOptions =
+        serde_json::from_str(&options.to_string()).unwrap();
+    
+    // TODO: !!! time input in ISO format.
+    // let input_string = &json_obj["input_string"].as_str().unwrap();
 
-    let formatter = match relative_time_formatter {
-        Ok(formatter) => formatter,
-        Err(error) => {
-            return Ok(json!({
-                "error": "Cannot create formatter",
-                "error_msg": error.to_string(),
-                "label": label,
-            }));
-        }
-    };
+    // let iso_input = DateTime::try_new_iso_datetime(input_string);
 
-    let formatted_result = formatter.format(count.clone());
-    let result_string = formatted_result.to_string();
+    let dtf = DateTimeFormatter::try_new(
+        &data_locale,
+        dt_options.into(),
+    )
+        .expect("Failed to create DateTimeFormatter instance.");
+
+    // !!! TEMPORARY !
+    let date_iso = DateTime::try_new_iso_datetime(2020, 9, 1, 12, 34, 28)
+        .expect("Failed to construct DateTime.");
+    let any_datetime = date_iso.to_any();
+
+    // !!! Calendar.
+    let calendar_type = gregorian;
+    let calendar_date = date_iso.to_calendar(calendar_type);
+
+    // Result to stdout.
+    // TODO: get the date/time info from a skeleton.
+    let formatted_dt = dtf.format(&any_datetime).expect("should work");
+    let result_string = formatted_dt.to_string();
+    
     Ok(json!({
         "label": label,
         "result": result_string,
-    "actual_options": format!("{options:?}"),
+        "actual_options": format!("{option_struct:?}, {options:?}, {locale_str:?}"),
     }))
+        
 }
